refactor: extract managed research run authority
